Simplify logic to find missing dependencies
This simplifies the construct introduced in
e160e92d8e2bf0ac36730913527a1b4a6353a7c0:
scalaz now no longer comes in transitively and
perhaps this logic doesn't justify introducing it.

Unfortunately we don't have a good automated test
for this feature yet, but I manually verified with
monix (which has some more interesting resolutions)
that it appears to behave the same.

Algorithm performance is not really a concern:
it is very fast in both implementations.