Support for local, nix-managed dependencies
This change extends the `sbtixBuildInputs` feature so that the added
dependencies can be used at `sbtix-gen` time, so transitive and overridden
dependencies are supported more accurately. This solution seems simpler
and also supports local dependencies that are not built with sbtix, as
long as they use a maven/ivy-style output directory layout and do not
rely on repositories not available in this build.

To see it in action, look at `tests/multi-build` or
https://github.com/raboof/sbtix-sample-local-build-inputs